Move existing vue components to pages folder. Created demo component
