+++ app/ui/src/components/linkGroup.vue
@@ -1,0 +1,32 @@
+<template>
+  <div class="card">
+    <header class="card-header">
+      <p class="card-header-title">
+        <slot name="header"></slot>
+      </p>
+    </header>
+    <div class="card-content">
+      <div class="content">
+        <slot></slot>
+      </div>
+    </div>
+    <footer class="card-footer">
+      <slot name="footer"></slot>
+    </footer>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'link-group',
+  data () {
+    return { }
+  },
+  methods: {
+  }
+}
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped>
+</style>
