+++ app/ui/src/pages/Main.vue
@@ -1,0 +1,22 @@
+<template>
+  <section class="section">
+    <p>{{ msg }}</p>
+  </section>
+</template>
+
+<script>
+export default {
+  name: 'Main',
+  data () {
+    return {
+      msg: 'Main Stub'
+    }
+  },
+  methods: {
+  }
+}
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped>
+</style>
