+++ app/ui/src/pages/Register.vue
