+++ app/ui/src/router/index.js
@@ -1,18 +1,19 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import HelloWorld from '@/components/HelloWorld'
-import Register from '@/components/Register'
-import LogIn from '@/components/LogIn'
-import Dashboard from '@/components/Dashboard'
+import Main from '@/pages/Main'
+import Register from '@/pages/Register'
+import LogIn from '@/pages/LogIn'
+import Dashboard from '@/pages/Dashboard'
+import MiningOperations from '@/pages/MiningOperations'
 
 Vue.use(Router)
 
 export default new Router({
   routes: [
-    // Example: ,{ path: '/', name: 'HelloWorld', component: HelloWorld }
-    { path: '/', name: 'HelloWorld', component: HelloWorld },
+    { path: '/', name: 'Main', component: Main },
     { path: '/Register', name: 'Register', component: Register },
     { path: '/LogIn', name: 'LogIn', component: LogIn },
-    { path: '/Dashboard', name: 'Dashboard', component: Dashboard }
+    { path: '/Dashboard', name: 'Dashboard', component: Dashboard },
+    { path: '/MiningOperations', name: 'MiningOperations', component: MiningOperations }
   ]
 })
